🔨 Move test util methods in a separate file
