+++ src/test/parser.test.ts
@@ -3,6 +3,7 @@ import { suite, test } from "mocha";
 import { CharmActions, CharmConfig, CharmToxConfig, MapWithNode, Problem, SequenceWithNode, WithNode } from "../model/charm";
 import { Range } from "../model/common";
 import { YAMLParser, parseCharmActionsYAML, parseCharmConfigYAML, parseCharmMetadataYAML, parseToxINI } from "../parser";
+import { newRange, unindent } from "./util";
 
 function cursorOverMap<T>(map: MapWithNode<T> | undefined) {
     let index = -1;
@@ -11,33 +12,6 @@ function cursorOverMap<T>(map: MapWithNode<T> | undefined) {
         next() { return entries[++index][1]; },
         get current() { return entries[index][1]; },
     };
-}
-
-function newRange(startLine: number, startCharacter: number, endLine: number, endCharacter: number): Range {
-    return {
-        start: { line: startLine, character: startCharacter },
-        end: { line: endLine, character: endCharacter },
-    };
-}
-
-function unindent(s: string): string {
-    const lines = s.split('\n');
-    if (lines[0] !== '') {
-        throw new Error('First line should be empty');
-    }
-    lines.splice(0, 1);
-
-    let indent = 0;
-    let index = 0;
-    const pattern = /^(\s+)/;
-    for (; index < lines.length; index++) {
-        const match = lines[index].match(pattern);
-        if (match) {
-            indent = match[1].length;
-            break;
-        }
-    }
-    return lines.map((x, i) => i >= index ? x.substring(indent) : x).join('\n').trimEnd();
 }
 
 suite(YAMLParser.name, function () {
